+++ src/effect.js
@@ -5,6 +5,7 @@ import type { StatePath } from "./storage";
 
 export type Effect<M> = {
   // FIXME: Remove srcPath once async replies is working
-  effect: (msg: M, srcPatH: StatePath) => ?Message | Promise<?Message>,
+  effect: (msg: M, srcPath: StatePath) => ?Message | Promise<?Message>,
+  name?: string,
   subscribe: Subscriptions<M>,
 };